requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2024.2.2
charset-normalizer==3.4.2
click==8.2.1
decorator==5.2.1
Flask==3.0.2
Flask-Cors==4.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpickle==4.1.0
jsonschema==4.24.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
neo4j==5.17.0
networkx==3.2.1
numpy==1.26.4
pandas==2.2.0
plotly==5.18.0
proto-plus==1.26.1
protobuf==4.25.8
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
requests==2.32.3
six==1.17.0
streamlit==1.32.0
tenacity==8.5.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2024.1
urllib3==2.4.0
Werkzeug==3.1.3
